+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change log
 
 ## [Unreleased]
+### Changed
+  * Remove loose mode usage of class properties/methods plugins.
+  * Core-js 3 upgrade.
+  * Change our targets and polyfill strategy.
+    * We want to leverage targets "intersection". 
+    * Drop IE11 support and use the `defaults` Browserslist query.
+    * Don't setup the Browserslist query inside the preset. 
+    * Use `useBuiltIns: 'usage'` for polyfills handling.
+    * Use React preset `useSpread` options.
+    * Use JSX automatic runtime.
 
 ## 5.13.1 - 2023-07-18
 ### Changed